Further work on new weapon +actions

+++ src/d_ticcmd.h
@@ -43,6 +43,6 @@ typedef struct
     signed char     forwardmove;    // * 2048 for move
     signed char     sidemove;       // * 2048 for move
     signed short    angleturn;      // << 16 for angle delta
-    byte            buttons;
+    int             buttons;
     int             lookdir;
 } ticcmd_t;
